Fix Sound and Save back to using Vector

+++ src/com/inspedio/enums/SaveDataType.java
@@ -26,7 +26,7 @@ public class SaveDataType extends InsEnum{
 		super(Name, Value);
 	}
 	
-	public static SaveDataType getType(int Value){
+	public static SaveDataType getType(short Value){
 		switch (Value){
 			case 0:
 				return BOOLEAN;
